Create a cabinet manager which can manage multiple cabinets.
Mainly, this is a store for cabinets to be registered with, that also
automatically handles ticking those cabinets in its own thread.
Basically, if you create a CabinetManager, you can add Cabinets to it
and it will handle sending games, polling, etc. You only need to ask
each cabinet for its status, or update the filename, and it will
automatically be handled for you.